cacher: move hasPathPrefix helper to new key subpackage
This way key.HasPathPrefix can be used from both cacher and store
