+++ staging/src/k8s.io/apiserver/pkg/storage/cacher/util_test.go
@@ -21,54 +21,6 @@ import (
 	"time"
 )
 
-func TestHasPathPrefix(t *testing.T) {
-	validTestcases := []struct {
-		s      string
-		prefix string
-	}{
-		// Exact matches
-		{"", ""},
-		{"a", "a"},
-		{"a/", "a/"},
-		{"a/../", "a/../"},
-
-		// Path prefix matches
-		{"a/b", "a"},
-		{"a/b", "a/"},
-		{"中文/", "中文"},
-	}
-	for i, tc := range validTestcases {
-		if !hasPathPrefix(tc.s, tc.prefix) {
-			t.Errorf(`%d: Expected hasPathPrefix("%s","%s") to be true`, i, tc.s, tc.prefix)
-		}
-	}
-
-	invalidTestcases := []struct {
-		s      string
-		prefix string
-	}{
-		// Mismatch
-		{"a", "b"},
-
-		// Dir requirement
-		{"a", "a/"},
-
-		// Prefix mismatch
-		{"ns2", "ns"},
-		{"ns2", "ns/"},
-		{"中文文", "中文"},
-
-		// Ensure no normalization is applied
-		{"a/c/../b/", "a/b/"},
-		{"a/", "a/b/.."},
-	}
-	for i, tc := range invalidTestcases {
-		if hasPathPrefix(tc.s, tc.prefix) {
-			t.Errorf(`%d: Expected hasPathPrefix("%s","%s") to be false`, i, tc.s, tc.prefix)
-		}
-	}
-}
-
 func TestCalculateRetryAfterForUnreadyCache(t *testing.T) {
 	tests := []struct {
 		downtime time.Duration
